% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-missing-columns.R
\name{add_missing_columns}
\alias{add_missing_columns}
\title{Add missing columns}
\usage{
add_missing_columns(df, full_column_list, fill_value = NA, reorder = TRUE)
}
\arguments{
\item{df}{A dataset}

\item{full_column_list}{A character vector of all the columns which should
appear in the dataset}

\item{fill_value}{The value that the new columns should be populated with}

\item{reorder}{Logical, default = \code{TRUE}; Should the columns of the output
data be reordered using the \code{full_column_list}?}
}
\description{
Given a list of columns which should occur in a dataset, adds in the missing
columns and fill them with the supplied default value. Will only add columns,
not remove them.
}
