% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-band-means.R
\name{calculate_riob}
\alias{calculate_riob}
\title{Calculate Robust Integration Over Brackets}
\usage{
calculate_riob(
  band_counts_df,
  bands_df,
  method = c("Geometric", "Arithmetic", "Harmonic", "Median"),
  min_alpha = 1.11
)
}
\arguments{
\item{band_counts_df}{Counts of bands, created using \code{\link[=compute_band_counts]{compute_band_counts()}};
fed into both \code{\link[=calculate_mcib]{calculate_mcib()}} and \code{\link[=calculate_rpme]{calculate_rpme()}}}

\item{bands_df}{A table of bands, created using \code{\link[=create_bands_df]{create_bands_df()}}; fed
into \code{\link[=calculate_mcib]{calculate_mcib()}}}

\item{method}{String; one of "Arithmetic", "Harmonic", "Geometric" or
"Median", fed into \code{\link[=calculate_rpme]{calculate_rpme()}}}

\item{min_alpha}{Numeric; the minimum threshold for the alpha parameter of
the Pareto distribution, fed into \code{\link[=calculate_rpme]{calculate_rpme()}}}
}
\description{
Used by \code{\link[=estimate_band_means]{estimate_band_means()}}. You can use this function to check that band
labels have been assigned correctly, otherwise use \code{\link[=estimate_band_means]{estimate_band_means()}}.
}
\seealso{
Other functions for estimating means of banded data:
\code{\link[=calculate_mcib]{calculate_mcib()}},
\code{\link[=calculate_rpme]{calculate_rpme()}},
\code{\link[=compute_band_counts]{compute_band_counts()}},
\code{\link[=create_bands_df]{create_bands_df()}},
\code{\link[=estimate_band_means]{estimate_band_means()}}
}
\author{
Farm Business Survey team (\href{mailto:fbs.queries@defra.gov.uk}{fbs.queries@defra.gov.uk})
}
\concept{functions for estimating means of banded data}
