% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-band-means.R
\name{calculate_rpme}
\alias{calculate_rpme}
\title{Robust Pareto Midpoint Estimator (RPME)}
\usage{
calculate_rpme(
  band_counts_df,
  method = c("Geometric", "Arithmetic", "Harmonic", "Median"),
  min_alpha = 1.11
)
}
\arguments{
\item{band_counts_df}{Counts of bands, created using \code{\link[=compute_band_counts]{compute_band_counts()}}}

\item{method}{What kind of average to return: "Arithmetic", "Harmonic",
"Geometric" or "Median"}

\item{min_alpha}{Numeric; the minimum threshold for the alpha parameter of
the Pareto distribution}
}
\description{
Used by \code{\link[=estimate_band_means]{estimate_band_means()}}. Run to check how RPME has been calculated.
}
\seealso{
Other functions for estimating means of banded data:
\code{\link[=calculate_mcib]{calculate_mcib()}},
\code{\link[=calculate_riob]{calculate_riob()}},
\code{\link[=compute_band_counts]{compute_band_counts()}},
\code{\link[=create_bands_df]{create_bands_df()}},
\code{\link[=estimate_band_means]{estimate_band_means()}}
}
\author{
Farm Business Survey team (\href{mailto:fbs.queries@defra.gov.uk}{fbs.queries@defra.gov.uk})
}
\concept{functions for estimating means of banded data}
