% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect-github-ssh.R
\name{connect_github_ssh}
\alias{connect_github_ssh}
\title{Connect RStudio to your GitHub account using SSH}
\usage{
connect_github_ssh(username, email)
}
\arguments{
\item{username}{user's GitHub username}

\item{email}{user's email address}
}
\value{
No return value. Side effects include setting Git credentials,
generating an SSH key, and establishing a GitHub connection.
}
\description{
Connect RStudio to your GitHub account using SSH - this is the
preferred method of connecting RStudio and GitHub on the DASH platform. The
user will only need to supply their GitHub username and email address, then
the function will set their credentials and generate an SSH key. Once the
SSH key has been added to GitHub, the function does the final bits of
set-up. All steps and outputs are printed in the RStudio console rather
than the terminal, which offers a more user friendly experience. This
method of connection is only recommended for the DASH platform - for
connecting local RStudio installs to GitHub, see
\code{\link[=connect_github_pat]{connect_github_pat()}} instead.
}
\details{
The suggested method for connecting RStudio and GitHub on the DASH
platform is via SSH. The alternative is via Personal Access Token, but this
method can be a bit clunky on the DASH platform. Connecting via SSH
provides a much smoother user experience.

This function is designed to simplify the connection process and avoid the
user having to interact with the terminal - which many beginners to Git
find intimidating. Instead, all prompts and outputs are returned in the
RStudio console.

The user calls the function and gives their user credentials (username and
email), the function will set these and print an output to confirm they
have been set. The function will then create a hidden SSH folder and
sub-folder named "id_ed25519" (~/.ssh/id_ed25519). Once done, it will then
generate your SSH key and save it to this folder. The SSH key will then be
printed in the console. The function then adds GitHub as a known host.

At this point, the function pauses and asks the user to confirm they have
added the SSH key to GitHub. The user can then copy the SSH key from the
console, go to GitHub -> settings -> SSH and GPG keys and add the SSH key.

Once added, the user can respond to the prompt and the function will finish
establishing the connection. If all works as expected, the message "Hi
username! You've successfully authenticated, but GitHub does not provide
shell access" will be printed in the console.

This method of connection is only recommended for the DASH platform.
}
\note{
This function is interactive, pausing for user input before continuing.
Therefore, this function is not suitable for automated scripts.
}
\examples{
\dontrun{
# Set GitHub credentials
connect_github_ssh(
  username = "my_github_username",
  email = "my_email"
)
}

}
\seealso{
\code{\link[=connect_github_pat]{connect_github_pat()}}
}
\author{
Josh Moatt
}
