% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-roxygen-script.R
\name{create_roxygen_script}
\alias{create_roxygen_script}
\title{Create a new script using the default template.}
\usage{
create_roxygen_script(
  file_name = NULL,
  file_path = NULL,
  author = NULL,
  email = NULL
)
}
\arguments{
\item{file_name}{string containing desired name for script.}

\item{file_path}{string containing folder name to save script. This is built
on the here function in R, so will follow your root directory. If you want
to save in a sub-folder, enter the full folder sequence, e.g.
"folder/sub-folder".}

\item{author}{string containing author's name.}

\item{email}{string containing author's email.}
}
\value{
An R script will be saved in the root directory or in the specified
folder.
}
\description{
This function will create a new R script with a roxygen2
template for use documenting functions. There are multiple options for
customisation (see details).
}
\details{
This function generates a new R script pre-filled with a standard
roxygen2 template.

You can optionally pre-fill the following fields:
\itemize{
\item Script name
\item Save location
\item Author name
\item Author email
}

The script is saved in your project's root directory (via \code{\link[here:here]{here::here()}})
or a specified folder.
}
\examples{
\dontrun{
# Create a script with default settings
create_roxygen_script()

# Create a script with custom metadata
create_roxygen_script(
  file_name = "my_function_name",
  file_path = "R",
  author = "Farming Stats",
  email = "Farming.Stats@defra.gov.uk"
)
}

}
\author{
Josh Moatt (\href{mailto:Joshua.Moatt@defra.gov.uk}{Joshua.Moatt@defra.gov.uk})
}
