% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-script.R
\name{create_script}
\alias{create_script}
\title{Create a new script using the default template.}
\usage{
create_script(
  file_name = NULL,
  file_path = NULL,
  author = NULL,
  email = NULL,
  date = format(Sys.Date(), "\%d/\%m/\%Y")
)
}
\arguments{
\item{file_name}{string containing desired name for script.}

\item{file_path}{string containing folder name to save script. This is built
on the here function in R, so will follow your root directory. If you want
to save in a sub-folder, enter the full folder sequence, e.g.
"folder/sub-folder".}

\item{author}{string containing author's name.}

\item{email}{string containing author's email.}

\item{date}{string containing a date. By default, this will be set as today's
date.}
}
\value{
An R script will be saved in the root directory or in the specified
folder.
}
\description{
This function will create a new R script using the default
header template. There are multiple options for customisation (see
details).
}
\details{
This function generates a new R script pre-filled with a standard
header template. By default, all fields are left blank except for the
creation date, which is set to today's date.

You can optionally pre-fill the following fields:
\itemize{
\item Script name
\item Save location
\item Author name
\item Author email
\item Date created
}

The script is saved in your project's root directory (via \code{\link[here:here]{here::here()}})
or a specified sub-folder.
}
\examples{
\dontrun{
# Create a script with default settings
create_script()

# Create a script with custom metadata
create_script(
  file_name = "analysis_script",
  file_path = "scripts",
  author = "Farming Stats",
  email = "Farming.Stats@defra.gov.uk"
)
}

}
\author{
Josh Moatt (\href{mailto:Joshua.Moatt@defra.gov.uk}{Joshua.Moatt@defra.gov.uk})
}
