% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defra-analysis-proj.R
\name{defra_analysis_proj}
\alias{defra_analysis_proj}
\title{Function for Defra analysis RStudio project template}
\usage{
defra_analysis_proj(path, ...)
}
\description{
This is a function that is called in the "New project" viewer pane when the
user chooses the Defra Analysis Project template. It should not be used away
from the RStudio "New project" viewer.

I have not included additional information on how to use this function, as it
is not intended to be used outside the template call.
}
\author{
Josh Moatt
}
