% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbs-round.R
\name{fbs_round}
\alias{fbs_round}
\title{Round numbers in the FBS style}
\usage{
fbs_round(x, round_to = NULL, value_type = c("", "area_gm"))
}
\arguments{
\item{x}{The number to round}

\item{round_to}{numeric, default = NULL; if NULL, will default to the FBS
style, alternatively, provide any power of 10 (including minus powers, e.g. 0.1)}

\item{value_type}{If "area_gm" is supplied, the different method of rounding
(explained above) will be used}
}
\value{
A rounded number in string format
}
\description{
The FBS style of rounding is show 0 as 0, rounding everything >= 100 to the
nearest 100, everything else that is >= 10 to the nearest 1, and everything
else to 1 decimal place
}
\details{
For areas and gross margins, rounding is slightly different: we show 0 as 0,
round everything else that is >= 10 to the nearest 1 and everything else to
1 decimal place

We can also use this function to round to any power of 10

Note that the janitor round_half_up function is used, as opposed to R's
default 'round-to-even' method
}
