% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-anon-workbook.R
\name{generate_anon_workbook}
\alias{generate_anon_workbook}
\title{Generate A Workbook Object From An 'aftable' without an author in the file}
\usage{
generate_anon_workbook(aftable)
}
\arguments{
\item{aftable}{An aftable-class object created using
\code{\link{create_aftable}} (or \code{\link{as_aftable}}), which
contains the data and information needed to create a workbook.}
}
\value{
A Workbook-class object
}
\description{
Populate an 'openxlsx' Workbook-class object with content from an
aftable-class object. In turn, the output can be passed to
\code{\link[openxlsx]{saveWorkbook}} from 'openxlsx'
}
\details{
Have added the \code{creator = NULL} argument into createWorkbook so that there
will be no file author
}
