% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-cell-style.R
\name{get_cell_style}
\alias{get_cell_style}
\title{Get cell styles for stats publications}
\usage{
get_cell_style(
  style1 = c("text", "number", "percent", "custom"),
  style2 = c("body", "heading", "highlight", "lastrow"),
  bold = FALSE,
  separator = FALSE,
  custom_format = NULL,
  indent = 0
)
}
\arguments{
\item{style1}{Whether the data is text, numbers or percentages. Text cells
get left-aligned, while number and percentage cells get right-aligned. By
default, both percentages and numbers are rounded to a whole number; to round
to a number of decimal places, define \code{style1} to \code{"custom"} and specify the
\code{custom_format} argument.}

\item{style2}{Where the data is in the table (heading row, body, average row
or last row). Heading cells are bold, aligned vertically in the middle, and
get upper and lower borders. All other cells are bottom-aligned. 'Highlight'
cells get upper and lower borders, while 'last row' cells just get lower borders.}

\item{bold}{logical, default = \code{FALSE}; should the text be bold?}

\item{separator}{logical, default = \code{FALSE}; should there be a separating
border to the right of the column?}

\item{custom_format}{Used by numFmt argument of \code{\link[openxlsx:createStyle]{createStyle()}}; an Excel
custom number format, e.g., "##0.0\%" rounds a percentage to 1 decimal place.}

\item{indent}{Numeric, default = 0; Horizontal indentation of cell contents.}
}
\description{
Get cell styles for stats publications
}
\author{
Farm Business Survey team (\href{mailto:fbs.queries@defra.gov.uk}{fbs.queries@defra.gov.uk})
}
