% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-ons-deflators.R
\name{get_gdp_deflators}
\alias{get_gdp_deflators}
\alias{get_ons_deflators}
\title{Read in ONS deflator data and run real terms calculation}
\usage{
get_gdp_deflators(deflator_years, index = c("GDP", "CPIH"))

get_ons_deflators(deflator_years, index = c("GDP", "CPIH"))
}
\arguments{
\item{deflator_years}{The years to get the ONS deflators for}

\item{index}{The index to use; either GDP or CPIH}
}
\value{
data.frame; The ONS deflators for all years in \code{deflator_years}
}
\description{
This mimics what is done for the Oct/Nov income publication
Works on the inbuilt GDP and CPIH data, which is downloaded from ONS
}
