% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-ons-series.R
\name{get_ons_series}
\alias{get_ons_series}
\title{Read in quarterly ONS price indices and convert to annual series}
\usage{
get_ons_series(
  series_years,
  index = c("", "GDP", "CPIH"),
  ons_url = NULL,
  get_snapshot = F,
  snapshot_date = NULL,
  save_path = NULL,
  year_end_q = 4
)
}
\arguments{
\item{series_years}{The years to get the ONS indices for (for financial
years, just the first year, e.g. 2020 for 2020/21)}

\item{index}{The index to use; either GDP or CPIH, or leave blank to download
another series (must provide a URL)}

\item{ons_url}{If provided, will download the CSV file at this URL to create
the indices; can be the full link to the inflation / price indices series,
just the URL extension, or the CSV generator (see details section)}

\item{get_snapshot}{Logical (default = \code{FALSE}); if \code{TRUE}, allows you to
choose a snapshot, rather than the live data}

\item{snapshot_date}{Optional date of the desired snapshot, must be a string
with format YYYY-MM-DD (i.e. coercible by \link[base:as.Date]{base::as.Date}); if not provided,
you will be prompted to choose a snapshot from a list of the latest ten}

\item{save_path}{If provided, will save the downloaded data to the specified
directory}

\item{year_end_q}{When converting the index from quarterly to annual, the
year-end to use, i.e. the last quarter of the year; default is 4, which
converts to calendar years (for financial years, set as 1)}
}
\value{
A tibble containing the ONS inflation / price indices series for all
years in \code{series_years} (for clarity, a column  of type \code{yearqtr} showing the
chosen year end is also included)
}
\description{
Retrieves an ONS inflation / price indices series. The URLs for the GDP and
CPIH series are included, or you can specify your own. Once the series has
been retrieved, you can convert a current terms value into real terms
using \code{value / index * 100}, where \code{index} is the price / deflator index for
the year the value corresponds to (see examples).
}
\details{
As an example, for the 'CPI ANNUAL RATE 00: ALL ITEMS 2015=100' series, the
following URL options will work:
\itemize{
\item https://www.ons.gov.uk/economy/inflationandpriceindices/timeseries/d7g7/mm23
\item economy/inflationandpriceindices/timeseries/d7g7/mm23
\item https://www.ons.gov.uk/generator?format=csv&uri=economy/inflationandpriceindices/timeseries/d7g7/mm23
}
}
\examples{
library(dplyr)
set.seed(1)

current_terms <- tibble(
  year = 2015:2020,
  prices = "Current",
  value = runif(n = 6, min = 100, max = 1000)
)

gdp_series <- get_ons_series(2015:2020, index = "GDP")

real_terms <- left_join(current_terms, gdp_series,
                        by = "year") \%>\%
  mutate(prices = "Real",
         value = value / index * 100)

bind_rows(current_terms, real_terms)

(current_terms$value / real_terms$index) * 100

}
