% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-dash-files.R
\name{list_dash_files}
\alias{list_dash_files}
\title{List files within a DASH directory}
\usage{
list_dash_files(path, max_tries = 5, interval = 2)
}
\arguments{
\item{path}{A string containing the path to the volume or folder. Should be
the full DASH string starting "/Volumes/..."}

\item{max_tries}{Maximum number of tries to access directory metadata.
Default is 5 attempts.}

\item{interval}{Interval between tries to read in data. Default is 2 seconds.}
}
\value{
List of files in DASH directory returned.
}
\description{
Lists files in a DASH directory (volume or folder) using the
\code{brickster} API. Designed to handle intermittent http2 errors by retrying
failed attempts. Requires the Defra DASH platform and appropriate
\code{brickster} environment variables.
}
\details{
This function wraps \code{\link[brickster:db_volume_list]{brickster::db_volume_list()}} with a retry loop
to handle intermittent http2 errors, which are common but typically
transient. These errors are not usually caused by incorrect code or file
paths, and rerunning the same request often succeeds.

The function will retry the request up to \code{max_tries} times, waiting
\code{interval} seconds between each attempt. If all attempts fail, an error
is thrown.

You must be working on the Defra DASH platform and have set the required
\code{brickster} environment variables, including a valid Databricks Personal
Access Token (PAT).
}
\examples{
\dontrun{
# list files in volume/directory
create_dash_dir(
  path = "/Volumes/prd_dash_lab/<volume-name>/<directory-name>"
)
}

}
\seealso{
\code{\link[brickster:db_volume_list]{brickster::db_volume_list()}}
}
\author{
Josh Moatt (\href{mailto:Joshua.Moatt@defra.gov.uk}{Joshua.Moatt@defra.gov.uk})
}
