% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-model-functions.R
\name{normalise_values}
\alias{normalise_values}
\title{Normalise values and check result}
\usage{
normalise_values(df, model_col, save_path = NULL, leave_one_out = TRUE, ...)
}
\arguments{
\item{df}{A dataframe with a column you are running a model on}

\item{model_col}{String; the name of the column you are running a model on}

\item{save_path}{The directory where the plot should be saved to}

\item{leave_one_out}{Passed to the \code{loo} argument of \code{\link[bestNormalize:bestNormalize]{bestNormalize::bestNormalize()}}}

\item{...}{Other arguments passed to \code{\link[bestNormalize:bestNormalize]{bestNormalize::bestNormalize()}}}
}
\description{
Uses \code{\link[bestNormalize:bestNormalize]{bestNormalize::bestNormalize()}} to select the best normalising
transformation for a vector of numeric values, then performs this
transformation. Will also creates a plot with two charts showing the
distributions of the untransformed and transformed variables, including the
name of the transformation used and the skewness value of the result. If
\code{save_path} is set, the plot will be saved as a PNG to the specified directory.
}
\seealso{
Other survey modelling functions:
\code{\link[=run_wald_test]{run_wald_test()}},
\code{\link[=summarise_variables]{summarise_variables()}},
\code{\link[=test_model_performance]{test_model_performance()}}
}
\author{
Farm Business Survey team (\href{mailto:fbs.queries@defra.gov.uk}{fbs.queries@defra.gov.uk})
}
\concept{survey modelling functions}
