% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ons-deflators.R
\docType{data}
\name{ons_deflators}
\alias{ons_deflators}
\alias{gdp_deflators}
\alias{cpih_deflators}
\title{Time series of the deflator data from ONS, updated every quarter}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 365 rows and 2 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 644 rows and 2 columns.
}
\source{
\url{https://www.ons.gov.uk/economy/grossdomesticproductgdp/timeseries/ybgb/}
}
\usage{
gdp_deflators

cpih_deflators
}
\description{
Time series of the deflator data from ONS, updated every quarter
}
\keyword{datasets}
