% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrite-num-cols.R
\name{overwrite_num_cols}
\alias{overwrite_num_cols}
\title{Fixes column where numbers are stored as text}
\usage{
overwrite_num_cols(excel_wb, sheet, cols, rows, df, num_format = "0")
}
\arguments{
\item{excel_wb}{Openxlsx workbook name}

\item{sheet}{Worksheet (either name as string, or location as numeric)}

\item{cols}{Vector of column numbers to be overwritten}

\item{rows}{Vector of row numbers to be overwritten}

\item{df}{Data frame containing the data from the relevant worksheet}

\item{num_format}{String, default = \code{"0"}; Excel-style cell format for the values}
}
\value{
Updated workbook with modified columns
}
\description{
If you have markers on your values (e.g. 'u' for low reliability, 'p' for
provisional, see \href{https://analysisfunction.civilservice.gov.uk/policy-store/symbols-in-tables-definitions-and-help/#section-6}{AF guidance}
for full list), you will still get 'numbers stored as text' errors on the
unmarked values, even with the latest version of \link[aftables:aftables-package]{aftables::aftables}, which uses \link[openxlsx2:openxlsx2-package]{openxlsx2::openxlsx2}.
}
\details{
This function overwrites only cells that can be converted to numbers (i.e. the
unmarked values), leaving the marked values and any suppressed values as-is.
This function works with both \link[openxlsx:openxlsx]{openxlsx::openxlsx} \code{Workbook} objects and \link[openxlsx2:openxlsx2-package]{openxlsx2::openxlsx2}
\code{wbWorkbook} objects.

This function is adapted from \code{rapid.spreadsheets::overwrite_df()}; for the
full rapid.spreadsheets code, see
\href{https://github.com/RAPID-ONS/rapid.spreadsheets/blob/main/R/create_data_table_tab.R}{the GitHub page}.
}
\examples{
\dontrun{
library(aftables)

set.seed(1)

# Create an aftable
cover_df <- list("Section" = c("Title", "Content"))

contents_df <- data.frame("Sheet name" = "Table",
                          "Sheet title" = "Example",
                          check.names = FALSE)

dummy_data <- round_with_commas(rnorm(10) * 1e5)

dummy_data_markers <- c(dummy_data[1:3],
                        paste(dummy_data[4], "[u]"),
                        dummy_data[5:10])

table_df <- data.frame(
  Category = LETTERS[1:10],
  "Dummy data" = dummy_data,
  "Dummy data with markers" = dummy_data_markers,
  check.names = FALSE)

aftable <- create_aftable(
  tab_titles = c("Cover", "Contents", contents_df$`Sheet name`),
  sheet_types = c("cover", "contents", "tables"),
  sheet_titles = c("Cover", "Contents", "Table"),
  sources = c(rep(NA_character_, 2), "Source"),
  tables = list(cover_df, contents_df, table_df))

excel_wb <- generate_workbook(aftable)

# Check the file
# note the format errors on the table sheet
openxlsx2::wb_open(excel_wb)

# Fix the errors
overwrite_num_cols(excel_wb, sheet = 3, cols = 2:3,
                   rows = 5:14, df = table_df)

# Check the file again
openxlsx2::wb_open(excel_wb)
}

}
\author{
Farm Business Survey team (\href{mailto:fbs.queries@defra.gov.uk}{fbs.queries@defra.gov.uk})
}
