% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-publication-files.R
\name{prepare_publication_files}
\alias{prepare_publication_files}
\title{Prepare publication files}
\usage{
prepare_publication_files(
  publication,
  target_dir,
  dev_mode = FALSE,
  overwrite = TRUE,
  open = TRUE
)
}
\arguments{
\item{publication}{String; The name of the publication in kebab-case, e.g.
"farm-accounts-england"}

\item{target_dir}{File path; Where to copy the publication files to. If using
the SCE, this should be a general folder where code is saved. If using
desktop, this should be a directory specific to the relevant publication run.}

\item{dev_mode}{Logical, default = FALSE; Are you in development mode (i.e.
working in a GitHub repo)?}

\item{overwrite}{Logical, default = TRUE; If files already exist, should they
be overwritten?}

\item{open}{Logical, default = TRUE; If \code{TRUE},
\link[usethis:proj_activate]{activates} the new project:
\itemize{
\item If using RStudio desktop, the package is opened in a new session.
\item If on RStudio server, the current RStudio project is activated.
\item Otherwise, the working directory and active project is changed.
}}
}
\value{
If open = TRUE, will open the newly created project (not recommended
in dev mode), otherwise will return a message stating whether the files have
been moved successfully or not
}
\description{
Copies publication files (run script and commentary scripts) to a specified
target directory. If you are using the SCE, will save them in a folder
with the same name as the \code{publication} argument. If you are on desktop,
will just save the files without putting them in a new folder.
}
