% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-dash-files.R
\name{read_file_from_volume}
\alias{read_file_from_volume}
\alias{read_csv_from_volume}
\alias{read_rds_from_volume}
\alias{read_xlsx_from_volume}
\alias{read_ods_from_volume}
\alias{read_rio_from_volume}
\alias{read_parquet_from_volume}
\title{Read data of various file types from the DASH data lake}
\usage{
read_file_from_volume(path, ext, max_tries = 5, interval = 2)

read_csv_from_volume(path, ..., max_tries = 5, interval = 2)

read_rds_from_volume(path, ..., max_tries = 5, interval = 2)

read_xlsx_from_volume(path, ..., max_tries = 5, interval = 2)

read_ods_from_volume(path, ..., max_tries = 5, interval = 2)

read_rio_from_volume(path, ..., max_tries = 5, interval = 2)

read_parquet_from_volume(path, ..., max_tries = 5, interval = 2)
}
\arguments{
\item{path}{A string containing the path to data to be read in. Should be the
full DASH string starting "/Volumes/..."}

\item{ext}{A string specifying the file type. Only used in the general
function. for the specific functions, this will be set to the correct
extension (e.g. .csv).}

\item{max_tries}{Maximum number of tries to read in data. Default is 5
attempts.}

\item{interval}{Interval between tries to read in data. Default is 2 seconds.}

\item{...}{Additional arguments passed to the appropriate reader function.}
}
\value{
A dataframe.
}
\description{
A set of convenience functions to read data from common file
formats from the DASH platform. Each function uses \code{brickster} and wraps
the appropriate reader function for each file type. These functions will
avoid the relatively frequent http2 errors that crop up with \code{brickster}.
See details for more info on how they solve this problem.  In order for
these functions to work, you must be working on the Defra DASH platform and
have set the required brickster environmental variables. There a specific
functions for reading in .csv, .xlsx, .Rds, and .ods files as well as a
generic function for reading in additional file types.
}
\details{
These functions are designed to handle the frequent http2 errors
that occur with \code{brickster}. From testing, these errors are not code or
file path errors, but are just minor bugs with the API. Often, if the same
code is rerun, the data will be read in fine.

There is one generic function which can be used to create a temporary file
for any file type (using the ext argument). There are also wrapper
functions for specifically reading in .csv, .Rds, and .xlsx files.

All of the functions work in the same way. They use
\code{\link[brickster:db_volume_read]{brickster::db_volume_read()}}, to attempt to read in the data and store it
in an appropriate temporary file. They use a retry loop that catch errors
with \code{\link[=tryCatch]{tryCatch()}}. If no error is thrown by \code{brickster}, the functions read
in the temporary file into R using an appropriate reader function and
returns the resulting data frame. If an error is thrown by \code{brickster}, the
functions will wait a set number of seconds (controlled by the interval
argument) before retrying. They will repeat the attempted data load until
either the file is read successfully or the maximum number of attempts (set
by max_tries) is reached - after which it throws an error.

In order for these functions to work, you must be working on the Defra DASH
platform and have set the required \code{brickster} environmental variables.
These variables include setting a databricks Personal Access Token (PAT).
For more information on how to do this, see the specific project README.

Reader functions used are as follows:

\itemize{
\item \strong{.csv} - \code{\link[readr:read_csv]{readr::read_csv()}}
\item \strong{.Rds} - \code{\link[readr:read_rds]{readr::read_rds()}}
\item \strong{.ods} - \code{\link[readODS:read_ods]{readODS::read_ods()}}
\item \strong{.xlsx (single worksheet)} - \code{\link[readxl:read_xlsx]{readxl::read_xlsx()}}
\item \strong{.xlsx (whole workbook)} - \code{\link[rio:import_list]{rio::import_list()}}
\item \strong{.parquet} - \link[arrow:read_parquet]{arrow::read_parquet}
}
}
\examples{
\dontrun{
# read Rds file
dat <- read_rds_from_volume(
  path = "/Volumes/prd_dash_lab/<path-to-file>/filename.Rds"
)

# read csv file
dat <- read_csv_from_volume(
  path = "/Volumes/prd_dash_lab/<path-to-file>/filename.csv",
  show_col_types = FALSE
)

# read ods file
dat <- read_ods_from_volume(
  path = "/Volumes/prd_dash_lab/<path-to-file>/filename.ods",
  sheet = "sheet-name"
)

# read xlsx single woksheet 
dat <- read_xlsx_from_volume(
  path = "/Volumes/prd_dash_lab/<path-to-file>/filename.xlsx",
  sheet = "sheet-name"
)

dat <- read_rio_from_volume(
  path = "/Volumes/prd_dash_lab/<path-to-file>/filename.xlsx"
)

dat <- read_parquet_from_volume(
  path = "/Volumes/prd_dash_lab/<path-to-file>/filename.parquet"
)
}

}
\seealso{
\code{\link[brickster:db_volume_read]{brickster::db_volume_read()}}, \code{\link[readr:read_csv]{readr::read_csv()}},
\code{\link[readr:read_rds]{readr::read_rds()}}, \code{\link[readxl:read_xlsx]{readxl::read_xlsx()}}, \code{\link[readODS:read_ods]{readODS::read_ods()}},
\code{\link[=read_csv_from_volume]{read_csv_from_volume()}},
\code{\link[=read_xlsx_from_volume]{read_xlsx_from_volume()}}, \code{\link[=read_rds_from_volume]{read_rds_from_volume()}}
\code{\link[rio:import_list]{rio::import_list()}}, \code{\link[=read_parquet_from_volume]{read_parquet_from_volume()}}
}
\author{
Josh Moatt (\href{mailto:Joshua.Moatt@defra.gov.uk}{Joshua.Moatt@defra.gov.uk})
}
