% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round-with-commas.R
\name{round_with_commas}
\alias{round_with_commas}
\title{Round numbers for publishing/sharing}
\usage{
round_with_commas(
  x,
  method = c("round_to", "optimise"),
  round_to = NULL,
  optimise_to = c("1", "100", "1k", "10k", "100k", "1m", "10m", "100m"),
  round_zeros = TRUE
)
}
\arguments{
\item{x}{The number to round}

\item{method}{Either \code{"round_to"} for rounding all numbers to a power of 10
(must supply \code{round_to} argument), or \code{"optimise"} for rounding smaller
numbers to more detail and larger numbers to less detail (see details)}

\item{round_to}{numeric; if method is set to \code{"round_to"}, provide any power
of 10 (including minus powers, e.g. 0.1)}

\item{optimise_to}{The maximum value values are rounded to, as a string
(1, or powers of 10 from 100 to 100 million, see details)}

\item{round_zeros}{When \code{method = "optimise"}, should zeros be shown with no
decimal places (\code{TRUE}), or to 1 decimal place (\code{FALSE})?}
}
\value{
A rounded value with comma separators
}
\description{
Round values with \code{\link[janitor:round_half_up]{janitor::round_half_up()}} (as opposed to the R default
'round-to-even' method) and add comma separators.
}
\details{
'Optimise' rounds smaller numbers to more detail and larger numbers to less
detail.

When \code{optimise_to} is set to \code{"1"}:
\itemize{
\item 0 is shown as 0
\item Values > 0 and < 10 are rounded to 1 decimal place
\item Values >= 10 are rounded to the nearest 1
}

When \code{optimise_to} is set to \code{"100"} the above plus:
\itemize{
\item Values >= 10 and < 100 are rounded to the nearest 1
\item Values >= 100 are rounded to the nearest 100
}

When \code{optimise_to} is set to \code{"1k"} (i.e. 1,000) the above plus:
\itemize{
\item Values >= 100 and < 1,000 are rounded to the nearest 100
\item Values >= 1,000 are rounded to the nearest 1,000
}

And so on, for \code{"10k"} (10,000), \code{"100k"} (100,000), \code{"1m"} (1,000,000),
\code{"10m"} (10,000,000), and \code{"100m"} (100,000,000),
}
