% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-model-functions.R
\name{run_wald_test}
\alias{run_wald_test}
\title{Wald test}
\usage{
run_wald_test(survey_model)
}
\arguments{
\item{survey_model}{A svyglm object}
}
\description{
Perform the Wald test on a survey model with \code{\link[survey:regTermTest]{survey::regTermTest()}}.
}
\seealso{
Tested on models created with \code{\link[survey:svyglm]{survey::svyglm()}}

Other survey modelling functions:
\code{\link[=normalise_values]{normalise_values()}},
\code{\link[=summarise_variables]{summarise_variables()}},
\code{\link[=test_model_performance]{test_model_performance()}}
}
\author{
Farm Business Survey team (\href{mailto:fbs.queries@defra.gov.uk}{fbs.queries@defra.gov.uk})
}
\concept{survey modelling functions}
