% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-databricks-pat.R
\name{set_databricks_pat}
\alias{set_databricks_pat}
\title{Set databricks Persoanl Access Token}
\usage{
set_databricks_pat()
}
\value{
prompt in R console to enter databricks PAT, which can then assigned
to a vector.
}
\description{
Simple function to set your databricks Personal Access Token
(PAT). Doing this via a function avoids the need to hardcode the PAT in a
script, which helps prevent accidentally committing it to GitHub. Not
needed on Posit Workbench, which handles credentials internally. But useful
for the RStudio server. Note, the function does not work in databricks
notebooks. Useful for when using the \code{brickster} package.
}
\examples{
\dontrun{
# set PAT
token <- set_databricks_pat()

# set to environmental variables
Sys.setenv(DATABRICKS_TOKEN = token)
}

}
\author{
Josh Moatt
}
