% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commentary-functions.R
\name{str_line_wrap}
\alias{str_line_wrap}
\title{Wrap strings to a set number of lines}
\usage{
str_line_wrap(string, lines)
}
\description{
Wrap strings to a set number of lines
}
\author{
Farm Business Survey team (\href{mailto:fbs.queries@defra.gov.uk}{fbs.queries@defra.gov.uk})
}
