% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-model-functions.R
\name{summarise_variables}
\alias{summarise_variables}
\title{Summarise variable to be modelled}
\usage{
summarise_variables(df, column)
}
\arguments{
\item{df}{A data set}

\item{column}{String; column to be summarised}
}
\description{
Summarise variable to be modelled
}
\seealso{
Other survey modelling functions:
\code{\link[=normalise_values]{normalise_values()}},
\code{\link[=run_wald_test]{run_wald_test()}},
\code{\link[=test_model_performance]{test_model_performance()}}
}
\author{
Farm Business Survey team (\href{mailto:fbs.queries@defra.gov.uk}{fbs.queries@defra.gov.uk})
}
\concept{survey modelling functions}
