% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uc-volume-get.R
\name{uc_volume_get}
\alias{uc_volume_get}
\title{Copy file from the DASH Unity Catalog}
\usage{
uc_volume_get(workspace, volume, token, out_file)
}
\arguments{
\item{workspace}{databricks workspace (string)}

\item{volume}{file path to data folder (string)}

\item{token}{databricks PAT (string)}

\item{out_file}{filepath on RStudio Cluster that data should be stored.}
}
\value{
data saved in Rstudio cluster.
}
\description{
This is a simple function to read data from the databricks
Unuity Catalog (UC) into RStudio. It is built using the \link[httr:GET]{httr::GET} function from
the httr package. This function was lifted directly from the DASH playbook.
For saving files, see \code{uc_volume_put}.

Note, you must have set up a databricks Personal Access Token (PAT) in the
databricks settings before doing this.
}
\examples{
\dontrun{
uc_volume_get(
  workspace = "https://adb-7422054397937474.14.azuredatabricks.net",
  volume = "/Volumes/prd_dash_lab/<path-to-file>/filename.csv",
  token = "dapid4b3d********************a687f9b",
  out_file = here::here("data", "filename.csv")
)
}

}
\author{
Josh Moatt
}
