% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uc-volume-put.R
\name{uc_volume_put}
\alias{uc_volume_put}
\title{Write file to the DASH Unity Catalog}
\usage{
uc_volume_put(workspace, volume, token, file, folder)
}
\arguments{
\item{workspace}{databricks workspace (string)}

\item{volume}{path to folder on databricks where data should be written
(string)}

\item{token}{databricks PAT (string)}

\item{file}{name of file to be exported including file extension. Will be
used to both specify the file to export and specify the file to be created
on databricks (String).}

\item{folder}{path to file you wish to save, excluding file name and
extension (string).}
}
\value{
file saved on DASH.
}
\description{
This is a simple function to save a file to the DASH Unity
Catalog (UC). It is built using the httr2 package and uses a PUT request to
save the file. It will save a local file, either from a drive or temp
directory. For reading data, see \code{uc_volume_get}.

Note, you must have set up a databricks Personal Access Token (PAT) in the
databricks settings before doing this.
}
\examples{
\dontrun{
uc_volume_put(
  workspace = "https://adb-7422054397937474.14.azuredatabricks.net",
  volume = "/Volumes/prd_dash_lab/<path-to-file>",
  token = "dapid4b3d********************a687f9b",
  file = "filename.csv",
  folder = here::here("data")
)
}

}
\author{
Josh Moatt
}
