% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-dash-files.R
\name{write_files_to_volume}
\alias{write_files_to_volume}
\alias{dash_volume_write}
\alias{write_xlsx_to_volume}
\alias{write_xlsx2_to_volume}
\alias{write_rds_to_volume}
\alias{write_csv_to_volume}
\alias{write_text_to_volume}
\alias{write_parquet_to_volume}
\title{Write files to the DASH Unity Catalog}
\usage{
dash_volume_write(..., max_tries = 5, interval = 2)

write_xlsx_to_volume(data, path, ...)

write_xlsx2_to_volume(data, path, ...)

write_rds_to_volume(data, path, ...)

write_csv_to_volume(data, path, ...)

write_text_to_volume(data, path, ...)

write_parquet_to_volume(data, path, ...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[brickster:db_volume_write]{brickster::db_volume_write()}}.}

\item{max_tries}{Maximum number of tries to read in data. Added to deal with
persistent http errors. Default is 5.}

\item{interval}{Interval between tries to read in data. Added to deal with
persistent http errors. Default is 2}

\item{data}{(All functions except \code{\link[=dash_volume_write]{dash_volume_write()}}) Object to be saved.
Usually this will be a data frame or similar. For \code{\link[=write_xlsx_to_volume]{write_xlsx_to_volume()}},
this must be a workbook object created using the \code{openxlsx} package.}

\item{path}{A string containing the path on DASH to save the file to. Should
be the full DASH string starting "/Volumes/..." including file name and
extension.}

\item{file}{(\code{\link[=dash_volume_write]{dash_volume_write()}} only) A string containing the local file
path to the file you want to save to DASH. Should be the full file path
including file name and extension.}
}
\value{
Data file saved to DASH volume.
}
\description{
Functions to streamline writing files to DASH volumes. These
functions will help avoid the http2 error which seems to be a frequent
problem with \code{brickster}. This suite of functions are built on
\code{\link[brickster:db_volume_write]{brickster::db_volume_write()}}. There is a general function,
\code{\link[=dash_volume_write]{dash_volume_write()}} which can be used to save previously
created files to the DASH platform. However, there are also more specific
functions, to save data frames a specific file types (.Rds, .csv, .xlsx
workbooks and rich text files like .md). These functions use the general function alongside an
appropriate write function, using a temporary file to avoid saving anything
locally. In order for these functions to work, you must be working on the
Defra DASH platform and have set the required \code{brickster} environmental
variables.
}
\details{
These functions are designed to handle the frequent http2 errors
that occur with \code{brickster}. From testing, these errors are not code or
file path errors, but are just minor bugs with the API. Often, if the same
code is rerun, the data will be saved without issue.

There is one generic function which can be used to write existing files to
DASH, covering any file type. There are also wrapper functions for
specifically writing .csv, .Rds, and .xlsx files as well as rich text files
like .txt, .md, .svg and .html.

All of the functions  use the general \code{\link[=dash_volume_write]{dash_volume_write()}}
function. This uses \code{\link[brickster:db_volume_write]{brickster::db_volume_write()}}, to attempt to write the
data to the specified DASH volume. It uses a retry loop that catches errors
with \code{\link[=tryCatch]{tryCatch()}}. If no error is thrown by \code{brickster}, the function will
end as the data has been written. If an error is thrown by \code{brickster}, the
function will wait a set number of seconds (controlled by the interval
argument) before retrying. It will repeat the attempt to save the data
until either the file is written successfully or the maximum number of
attempts (set by max_tries) is reached - after which it throws an error.

The more specific functions will save objects from your R environment
(usually data frames) to DASH directly, without the need to save them
locally first. This helps avoiding saving data to the project folder, thus
reducing the risk of accidentally committing the data or output to GitHub.
To do this, the functions use an appropriate write function (see below) to
create a temporary file of the right format (e.g. .csv). This file is then
passed to the general function \code{\link[=dash_volume_write]{dash_volume_write()}} and the
file is written to DASH as described above.

Reader functions used are as follows:

\itemize{
\item \strong{.csv} - \code{\link[readr:write_delim]{readr::write_csv()}}
\item \strong{.xlsx} - \code{\link[openxlsx:saveWorkbook]{openxlsx::saveWorkbook()}}
\item \strong{.xlsx} - \code{\link[openxlsx2:wb_save]{openxlsx2::wb_save()}}
\item \strong{.Rds} - \code{\link[base:readRDS]{base::saveRDS()}}
\item \strong{.md} - \code{\link[base:cat]{base::cat()}}
\item \strong{.parquet} - \code{\link[arrow:write_parquet]{arrow::write_parquet()}}
}

In order for these functions to work, you must be working on the Defra DASH
platform and have set the required brickster environmental variables. These
variables include setting a databricks Personal Access Token (PAT). For
more information on how to do this, see the specific project README.
}
\examples{
\dontrun{
# write existing file
dash_volume_write(
  path = "/Volumes/prd_dash_lab/<path-to-file>/filename.docx",
  file = here::here("outputs", "filename.docx")
)

# write csv file
write_csv_to_volume(
  path = "/Volumes/prd_dash_lab/<path-to-file>/filename.csv",
  data = my_data_frame
)

# write xlsx file (using openxlsx)
write_xlsx_to_volume(
  path = "/Volumes/prd_dash_lab/<path-to-file>/filename.xlsx",
  data = my_data_frame
)

# write xlsx file (using openxlsx2)
write_xlsx2_to_volume(
  path = "/Volumes/prd_dash_lab/<path-to-file>/filename.xlsx",
  data = my_data_frame
)

# write Rds file
write_rds_to_volume(
  path = "/Volumes/prd_dash_lab/<path-to-file>/filename.Rds",
  data = my_data_frame
)

# write markdown file
write_text_to_volume(
  path = "/Volumes/prd_dash_lab/<path-to-file>/filename.md",
  data = my_data
)

# write markdown file
write_parquet_to_volume(
  path = "/Volumes/prd_dash_lab/<path-to-file>/filename.parquet",
  data = my_data
)
}

}
\seealso{
\code{\link[brickster:db_volume_write]{brickster::db_volume_write()}}, \code{\link[readr:write_delim]{readr::write_csv()}},
\code{\link[openxlsx:saveWorkbook]{openxlsx::saveWorkbook()}}, \code{\link[base:readRDS]{base::saveRDS()}}, \code{\link[openxlsx2:wb_save]{openxlsx2::wb_save()}},
\code{\link[=write_xlsx_to_volume]{write_xlsx_to_volume()}}, \code{\link[=write_csv_to_volume]{write_csv_to_volume()}},
\code{\link[=write_rds_to_volume]{write_rds_to_volume()}},
\code{\link[=write_parquet_to_volume]{write_parquet_to_volume()}}
}
\author{
Josh Moatt (\href{mailto:Joshua.Moatt@defra.gov.uk}{Joshua.Moatt@defra.gov.uk})

Tom Pearson (\href{mailto:Thomas.Pearson@defra.gov.uk}{Thomas.Pearson@defra.gov.uk})
}
