% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-band-means.R
\name{calculate_mcib}
\alias{calculate_mcib}
\title{Mean Constrained Integration Over Brackets (MCIB)}
\usage{
calculate_mcib(band_counts_df, bands_df)
}
\arguments{
\item{band_counts_df}{Counts of bands, created using \code{\link[=compute_band_counts]{compute_band_counts()}}}

\item{bands_df}{A table of bands, created using \code{\link[=create_bands_df]{create_bands_df()}}}
}
\description{
Used by \code{\link[=estimate_band_means]{estimate_band_means()}}. Run to check how MCIB has been calculated.
}
\details{
\url{https://journals.sagepub.com/doi/10.1177/0081175018782579#sec-4}
}
\seealso{
Other functions for estimating means of banded data: 
\code{\link{calculate_riob}()},
\code{\link{calculate_rpme}()},
\code{\link{compute_band_counts}()},
\code{\link{create_bands_df}()},
\code{\link{estimate_band_means}()}
}
\author{
Farm Business Survey team (\href{mailto:fbs.queries@defra.gov.uk}{fbs.queries@defra.gov.uk})
}
\concept{functions for estimating means of banded data}
