% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commentary-functions.R
\name{compare_two_years}
\alias{compare_two_years}
\title{Compare two year's worth of data}
\usage{
compare_two_years(
  df,
  year_1,
  year_2,
  year_col = "survey_year",
  grouping_col = "grouping_factor",
  group_col = "group",
  group_levels = NULL,
  group_labels = group_levels,
  value_col = "value",
  price_type = NULL,
  extra_vars = NULL,
  diff = "percent",
  round_to = NULL,
  prefix = ""
)
}
\arguments{
\item{df}{A data frame}

\item{year_1}{First year to include in the comparison}

\item{year_2}{Second year to include in the comparison}

\item{year_col}{Default = \code{"survey_year"}; The column with the years in}

\item{grouping_col}{Default = \code{NULL}; The column with the grouping variables
in - only required if you want this column returned}

\item{group_col}{Default = \code{NULL}; The column with the group variables in -
only required if you want this column returned}

\item{group_levels}{Default = \code{NULL}; If you have a 'group' column that you are
using in the commentary, you may want to edit the levels of this column so that
they read better in text, e.g. "LFA grazing livestock" rather than "Grazing
livestock (LFA)"; to do this, set the current levels of the group column here,
and the labels you want with \code{group_labels}}

\item{group_labels}{Default = \code{group_levels}; The new level labels for the group column}

\item{value_col}{The column with the values in}

\item{price_type}{Default = \code{NULL}; If the data contains both real and
current prices, which should be included? To include both keep as \code{NULL}.
This requires the data to contain a column called 'prices'.}

\item{extra_vars}{String; Extra variable(s) to select, e.g. band}

\item{diff}{String; Either \code{"percent"} (default; for difference in \%) or
\code{"points"} (for difference in percentage points)}

\item{round_to}{passed to \code{\link[=fbs_round]{fbs_round()}}}

\item{prefix}{Default = \code{""}; If you are calculating percentage change, you
can add a prefix to the returned values, e.g. \code{"£"}}
}
\description{
Data must contain the following four columns: survey_year, grouping_factor,
group, value
}
\seealso{
Other commentary functions: 
\code{\link{create_list}()},
\code{\link{get_address}()},
\code{\link{get_diff_in_words}()},
\code{\link{get_pc_in_words}()}
}
\concept{commentary functions}
