% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-band-means.R
\name{compute_band_counts}
\alias{compute_band_counts}
\title{Count the population in each band using weights}
\usage{
compute_band_counts(
  df,
  bands_df,
  band_col,
  grouping_cols = "year",
  weights = "weight"
)
}
\arguments{
\item{df}{A data frame containing banded incomes}

\item{bands_df}{A table of bands, corresponding with the values in \code{band_col},
created using \code{\link[=create_bands_df]{create_bands_df()}}}

\item{band_col}{String; The name of the column in \code{df} with bands in}

\item{grouping_cols}{String (default = "year"); Columns to group the counts
by, along with \code{band} (if no groups, set to \code{NULL})}

\item{weights}{String (default = "weight"); The name of the column containing
the weights}
}
\description{
Used by \code{\link[=estimate_band_means]{estimate_band_means()}}
}
\seealso{
Other functions for estimating means of banded data: 
\code{\link{calculate_mcib}()},
\code{\link{calculate_riob}()},
\code{\link{calculate_rpme}()},
\code{\link{create_bands_df}()},
\code{\link{estimate_band_means}()}
}
\author{
Farm Business Survey team (\href{mailto:fbs.queries@defra.gov.uk}{fbs.queries@defra.gov.uk})
}
\concept{functions for estimating means of banded data}
