% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect-github-pat.R
\name{connect_github_pat}
\alias{connect_github_pat}
\title{Connect RStudio to your GitHub account using a Personal Access Token}
\usage{
connect_github_pat(username, email)
}
\arguments{
\item{username}{string containing GitHub username}

\item{email}{string containing email address used for gitHub account}
}
\value{
GitHub credentials and PAT set
}
\description{
This is a simple function which connect RStudio to GitHub using
a Personal Access Token, allowing you to work on GitHub repositories. This
function is for local RStudio installs only, not for on the DASH platfrom.
For the dash platform, use \code{connect_github_ssh} instead.
}
\details{
This function will set you GitHub credentials and Personal Access
Token (PAT) and connect RStudio to GitHub. This is essential if you want to
work in RStudio in projects/repos stored on GitHub.

It uses \code{\link[=system]{system()}} to run the necessary code in the terminal to set your
credentials, and \code{\link[gitcreds:gitcreds_get]{gitcreds::gitcreds_set()}} and \code{\link[credentials:set_github_pat]{credentials::set_github_pat()}} from the
\code{gitcreds} and \code{credentials} packages to connect your RStudio to GitHub.

\code{\link[gitcreds:gitcreds_get]{gitcreds::gitcreds_set()}} is an interactive function and will prompt users for
input. To replace existing credentials/PAT choose option 2. You will then
be prompted for you PAT. PAT should be changed every 30 days to ensure
security.

Note: For this function to work you must:

\itemize{
\item have git installed on your local machine
\item have a GitHub account
\item have created a Personal Access Token (PAT) on GitHub.
}

An additional feature I have added, not mentioned in the Defra instructions
is to add the \code{\link[credentials:set_github_pat]{credentials::set_github_pat()}} function call to an .Rprofile. This will
ensure your PAT is set for every R session, meaning you wont need to
provide your PAT when running functions such as \code{\link[xfun:install_github]{xfun::install_github()}} from the
devtools package. The function will check if a .Rprofile file already
exists, if one does, it will add the code to the bottom of the existing
profile. If the .Rprofile file does not exist, the function will create one
and add the code to it.

This function is for local RStudio installs only, not for on the DASH
platform.
}
\examples{
\dontrun{
# Set GitHub credentials
connect_github_pat(
  username = "my_github_username",
  email = "my_email"
)
}

}
\author{
Josh Moatt
}
