% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-band-means.R
\name{create_bands_df}
\alias{create_bands_df}
\title{Create bands}
\usage{
create_bands_df(
  start = 0,
  upper_limits = NULL,
  open_ended = FALSE,
  codes = NULL,
  prefix = ""
)
}
\arguments{
\item{start}{Numeric (default = 0); the start of the range}

\item{upper_limits}{Numeric (default = \code{NULL}); the upper limits of each
band, only used if \code{step} is not specified}

\item{open_ended}{Logical (default = \code{FALSE}); whether to add a open-ended
band as the last band (i.e. include all values greater than the upper limit)}

\item{codes}{Numeric (default = \code{NULL}); the codes used in the dataset for each
band - if NULL, will use a numeric sequence from 1 to the total number of bands}

\item{prefix}{String; a prefix to apply to the band values (for example, "£"
when the values are in GBP)}
}
\value{
A tibble with columns for the band codes, lower limits, upper limits
and band labels
}
\description{
Creates a tibble of bands with corresponding lower and upper limits and labels.
}
\examples{
create_bands_df(start = 1, upper_limits = c(10, 1000, 10000), prefix = "£")
create_bands_df(start = 0, upper_limits = c(10, 1000, 10000), open_ended = TRUE)

}
\seealso{
Other functions for estimating means of banded data: 
\code{\link{calculate_mcib}()},
\code{\link{calculate_riob}()},
\code{\link{calculate_rpme}()},
\code{\link{compute_band_counts}()},
\code{\link{estimate_band_means}()}
}
\author{
Farm Business Survey team (\href{mailto:fbs.queries@defra.gov.uk}{fbs.queries@defra.gov.uk})
}
\concept{functions for estimating means of banded data}
