% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-defra-stats-infographic.R
\name{create_defra_stats_infographic}
\alias{create_defra_stats_infographic}
\title{Create stats infographic for sharing to Defra stats X account}
\usage{
create_defra_stats_infographic(title, subtitle, source, key_points, icons)
}
\arguments{
\item{title}{The filepath to find the icons listed in the \code{icons} argument}

\item{subtitle}{The filepath to find the icons listed in the \code{icons} argument}

\item{source}{The source of the data within the infographic}

\item{key_points}{List of key points (use markdown formatting); between 3 and
5 points works best}

\item{icons}{List of icons; the icons will go alongside each key point (for a
list of the available icons, run \code{\link[=view_available_icons]{view_available_icons()}})}
}
\description{
Uses \link{grid} to output an infographic into the Plots pane. To save it, click
'Export', choose 'Save as Image...' and save as a PNG file. Adjust the height
and width to suit. NB: this function could use \code{\link[grDevices:png]{grDevices::png()}} rather than
manually exporting, but testing this has resulted in corrupted files. If you
can get it to work (including the option to edit height and width), please
feel free to submit your fix in a pull request.
}
\details{
Once you have saved your infographic, send it to the user engagement team.
Include:
\itemize{
\item The name of the publication
\item A link to the publication
\item The infographic .png file
\item Plain text of the content of the infographic, which will be used in the alt
text of the image
}

For example:

Hi User Engagement team, my team have just published 'Energy use on farms in
England, 2023/24'. Please find the infographic attached. The text within the
infographic reads:

Energy use on farms in England, 2023/24
\itemize{
\item Red diesel was used by \strong{98\%} of farms
\item \strong{32\% generate renewable energy} (mostly solar power)
\item \strong{84\%} of solar-generating farm businesses have panels installed on
farm building rooftops
\item \strong{20\%} of farms had conducted a \strong{carbon audit}
}

Source: Defra, Farm Business Survey England 2023/24
}
\examples{
title <- "Energy use on farms in England"
subtitle <- "2023/24"
source <- "Source: Defra, Farm Business Survey England 2023/24"
icons <- c("tractor", "sun", "barn", "form")
key_points <- list(
  "Red diesel was used<br>by **98\%** of farms",
  "**32\% generate renewable energy**<br>(mostly solar power)",
  "**84\%** of solar-generating farm<br>businesses have panels installed on<br>farm building rooftops",
  "**20\%** of farms had<br>conducted a **carbon audit**")

create_defra_stats_infographic(title, subtitle, source, key_points, icons)

}
\author{
Farm Business Survey team (\href{mailto:fbs.queries@defra.gov.uk}{fbs.queries@defra.gov.uk})
}
