% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commentary-functions.R
\name{create_list}
\alias{create_list}
\title{Create a list to insert into markdown text}
\usage{
create_list(
  string_vector,
  oxford_comma = FALSE,
  descriptor = NULL,
  last_connector = "and",
  case = c("original", "lower", "upper", "sentence", "title")
)
}
\arguments{
\item{string_vector}{A string vector}

\item{oxford_comma}{Logical (default = \code{FALSE}); should the last item be
connected using an Oxford comma?}

\item{descriptor}{A descriptor to add to the end of the list; if supplied,
provide an object of length 1 (if always should be the same word), 2 (if you
just require singular and plural options - in that order), or 3 (if you
require options for one, two or more than two items - in that order)}

\item{last_connector}{What word should connect the last two items? Default = \code{"and"}}

\item{case}{Option to change the case of the output using \code{\link[stringr:case]{stringr::case()}}}
}
\description{
Takes a vector of strings and collapses them into a list; the
last item is connected to the list with 'and', the rest of the items are
separated by commas
}
\examples{
create_list(string_vector = c("apple", "banana", "pear"), descriptor = c("fruit", "fruits"))
create_list(string_vector = c("apple", "banana"), descriptor = c("fruit", "fruits"))
create_list(string_vector = c("apple", "banana", "pear"), oxford_comma = T, descriptor = c("fruit", "fruits"))

}
\seealso{
Other commentary functions: 
\code{\link{compare_two_years}()},
\code{\link{get_address}()},
\code{\link{get_diff_in_words}()},
\code{\link{get_pc_in_words}()}
}
\author{
Farm Business Survey team (\href{mailto:fbs.queries@defra.gov.uk}{fbs.queries@defra.gov.uk})
}
\concept{commentary functions}
