% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commentary-functions.R
\name{cut_long_scale}
\alias{cut_long_scale}
\title{Adapt scales::cut_long_scale to allow lower case}
\usage{
cut_long_scale(lower = T, space = FALSE)
}
\description{
Adapt scales::cut_long_scale to allow lower case
}
\seealso{
\code{\link[scales:number]{scales::cut_long_scale()}}
}
