% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode-multi-choice.R
\name{decode_multi_choice_column}
\alias{decode_multi_choice_column}
\title{Decode a multiple choice column}
\usage{
decode_multi_choice_column(input_data, var, id_cols)
}
\arguments{
\item{input_data}{Dataset}

\item{var}{String; Column to decode}

\item{id_cols}{String; All columns used to uniquely identify the row}
}
\description{
When a survey, such as the Farm Business Survey, requires all answers to be
input as numbers, a geometric sequence (1, 2, 4, 8, 16, etc.) is used to code
multiple choice answers.
For example, the respondent choosing the first (code 1) and third (code 4)
options would be input as 5.
This method means that any combination of choices can be decoded into its
constituent parts.
This function decodes these answers by taking an input dataset with a column
of values and returning separate columns for each answer.
}
\seealso{
Other functions for decoding multiple choice answers: 
\code{\link{decompose_multi_choice_column}()},
\code{\link{decompose_multi_choice_value}()}
}
\concept{functions for decoding multiple choice answers}
