% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode-multi-choice.R
\name{decompose_multi_choice_value}
\alias{decompose_multi_choice_value}
\title{Decompose multiple choice value}
\usage{
decompose_multi_choice_value(n)
}
\arguments{
\item{n}{Value to decompose}
}
\description{
When a survey, such as the Farm Business Survey, requires all answers to be
input as numbers, a geometric sequence (1, 2, 4, 8, 16, etc.) is used to code
multiple choice answers.
For example, the respondent choosing the first (code 1) and third (code 4)
options would be input as 5.
This method means that any combination of choices can be decoded into its
constituent parts.
This function decodes these answers by taking an input value and
returning separate values for each answer.
}
\seealso{
Other functions for decoding multiple choice answers: 
\code{\link{decode_multi_choice_column}()},
\code{\link{decompose_multi_choice_column}()}
}
\author{
Farm Business Survey team (\href{mailto:fbs.queries@defra.gov.uk}{fbs.queries@defra.gov.uk})
}
\concept{functions for decoding multiple choice answers}
