% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy-plot.R
\name{easy_plot}
\alias{easy_plot}
\title{Use ggplot and afcharts to create accessible static charts}
\usage{
easy_plot(
  plot_data,
  aesthetics,
  chart_type = c("stacked", "grouped", "line", "distribution", "boxplot"),
  chart_direction = c("vertical", "horizontal"),
  labels = F,
  label_size = 8,
  label_colour = "white",
  label_position = position_stack(0.5),
  error_bars = F,
  series_breaks = NULL,
  x_axis_title = NULL,
  y_axis_title = NULL,
  y_label_function = scales::label_comma(),
  y_axis_breaks = 8,
  y_min = 0,
  y_max = NA,
  wrap_vars = NULL,
  wrap_type = c("axis", "plot"),
  wrap_cols = NULL,
  wrap_rows = NULL,
  wrap_char_width = 25,
  font_family = "GDS Transport Website",
  font_size = 24,
  af_palette = names(afcharts::af_colour_palettes),
  custom_palette = NULL,
  gridlines = NULL,
  zero_line = TRUE,
  legend_position = NULL,
  legend_justification = NULL,
  legend_cols = NULL,
  annotate_lines = TRUE,
  top_margin = 10,
  right_margin = 0,
  bottom_margin = 0,
  left_margin = 0
)
}
\arguments{
\item{plot_data}{Data to plot}

\item{aesthetics}{Mapping aesthetics using \code{\link[ggplot2:aes]{ggplot2::aes()}}; if you want a
line chart with a single group, remember to include \code{group = 1}}

\item{chart_type}{One of "stacked", "grouped", "line", "distribution" or
"boxplot": for a bar chart with only one group, both "stacked" and "grouped"
will work ("stacked" is the default); for boxplots, if you supply \code{ymin},
\code{lower}, \code{middle}, \code{upper} and \code{ymax} to the aesthetics, \code{\link[ggplot2:geom_boxplot]{ggplot2::geom_boxplot()}}
will use \code{stat = "identity"}, otherwise it will use \code{stat = "boxplot"}}

\item{chart_direction}{One of "vertical" or "horizontal" ("horizontal" uses
\code{\link[ggplot2:coord_flip]{ggplot2::coord_flip()}}; keep in mind that this will, by design, reverse the
order of items on the x-axis - this can be fixed using \code{\link[=rev]{rev()}} in \code{aesthetics}
or by reversing the factor levels in \code{plot_data})}

\item{labels}{If TRUE, adds \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}} labels}

\item{label_size}{Default = 8; Font size of labels}

\item{label_colour}{Default = white; Colour of labels}

\item{label_position}{Default = \code{position_stack(0.5)} for labelling stacked
bar charts; Can take a ggplot \verb{position_*} or a length 2 vector referring to
horizontal and vertical justification, respectively, e.g. \code{c(0, 0.5)} or
\code{c(NULL, 1)}}

\item{error_bars}{If TRUE, adds confidence intervals; for bar charts uses
\code{\link[ggplot2:geom_linerange]{ggplot2::geom_errorbar()}}, and for line charts uses \code{\link[ggplot2:geom_ribbon]{ggplot2::geom_ribbon()}}}

\item{series_breaks}{If supplied, will use the xintercept argument of
\code{\link[ggplot2:geom_abline]{ggplot2::geom_vline()}} to add breaks to the x-axis at the specified points;
for line charts, remember to set the \code{group} argument in \code{\link[ggplot2:aes]{ggplot2::aes()}} to ensure
the lines break properly (for a single group, use the column containing the
series names, or for multiple groups, an \code{\link[=interaction]{interaction()}} between the grouping
column and the series column)}

\item{x_axis_title}{If supplied, adds a title to the x-axis}

\item{y_axis_title}{If supplied, adds a title to the y-axis}

\item{y_label_function}{Default = \code{\link[scales:label_number]{scales::label_comma()}}; a function which
is applied to the y-axis labels within the \code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}}
function (set to \code{NULL} to return the default labels)}

\item{y_axis_breaks}{Default = 8; the number of breaks to show on the y-axis}

\item{y_min}{Default = 0; The minimum value of the y-axis (must be either 0
or below 0)}

\item{y_max}{Default = NA; The maximum value of the y-axis}

\item{wrap_vars}{If supplied, will feed into the facets argument of
\code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}}

\item{wrap_type}{One of "axis" (to return one plot with a wrapped axis) or
"plot" (to return multiple plots)}

\item{wrap_cols}{Feeds into the ncol argument of \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}};
only used with \code{wrap_type = "plot"}}

\item{wrap_rows}{Feeds into the nrow argument of \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}};
only used with \code{wrap_type = "plot"}}

\item{wrap_char_width}{Default = 25; Wraps facet labels to the specifies
number of characters}

\item{font_family}{Default = "GDS Transport Website"; This font must be
loaded for it to be used, and should only be used when publishing on gov.uk}

\item{font_size}{Default = 24; Font size for all chart text (except bar labels)}

\item{af_palette}{Which one of the \link[afcharts:af_colour_palettes]{afcharts::af_colour_palettes} to use
(default = \code{"main"}); for boxplots, the dark colours do not have enough
contrast with the boxplot lines so these are filtered out, giving a maximum
of three colours available}

\item{custom_palette}{A vector of colours; If supplied, will use this rather
than any of the AF palettes (useful if you need to reorder an AF palette or
need more than 6 colours)}

\item{gridlines}{Fed into the grid argument of \code{\link[afcharts:theme_af]{afcharts::theme_af()}}; If
supplied, will override defaults (i.e. "y" for vertical bar charts, "x"
for horizontal bar charts, and "xy" for line charts)}

\item{zero_line}{Default = TRUE; If set to FALSE, will remove origin line
(only do this if your y-axis does not include 0)}

\item{legend_position}{If supplied, will override defaults (i.e. "right"
for vertical charts and "top" for horizontal charts)}

\item{legend_justification}{If supplied, will override defaults (i.e. "top"
for vertical charts and "right" for horizontal charts)}

\item{legend_cols}{Fed into the ncol argument of \code{\link[ggplot2:guide_legend]{ggplot2::guide_legend()}}}

\item{annotate_lines}{Uses \code{\link[ggrepel:geom_text_repel]{ggrepel::geom_text_repel()}} to annotate lines in
line charts, instead of using a legend; set to \code{FALSE} to use a legend
instead, noting that it is more accessible if the resulting legend is in the
order of the lines (see examples for how to do this)}

\item{top_margin}{Default = 10; Top margin of plot, fed into the plot.margin
argument of \code{\link[ggplot2:theme]{ggplot2::theme()}} using \code{\link[ggplot2:element]{ggplot2::margin()}}}

\item{right_margin}{Default = 0; Right margin of plot}

\item{bottom_margin}{Default = 0; Bottom margin of plot}

\item{left_margin}{Default = 0; Left margin of plot}
}
\description{
Has a large number of options so that the need to modify using ggplot code
is minimised. Requires the y-axis to be continuous, but the x-axis can be of
any scale.
}
\examples{
# Creating a line chart without annotating the lines
# but correcting the legend order

test_df <- dplyr::tibble(
  letters = rep(LETTERS[1:4], 3),
  grouping = c(rep("1", 4), rep("2", 4), rep("3", 4)),
  numbers = c(seq(1000, 4000, 1000),
              seq(1250, 4250, 1000),
              seq(1500, 4500, 1000))
)

var_order <- dplyr::filter(test_df, letters == max(letters)) |>
  dplyr::arrange(desc(numbers)) |>
  dplyr::pull(grouping)

plot_df <- test_df |>
  dplyr::mutate(grouping = factor(grouping, levels = var_order))

easy_plot(plot_df, aes(x = letters, y = numbers,
                       colour = grouping, group = grouping),
          chart_type = "line", annotate_lines = FALSE)

}
\author{
Farm Business Survey team (\href{mailto:fbs.queries@defra.gov.uk}{fbs.queries@defra.gov.uk})
}
