% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-band-means.R
\name{estimate_band_means}
\alias{estimate_band_means}
\title{Apply Robust Integration Over Brackets}
\usage{
estimate_band_means(
  df,
  bands_df,
  band_col,
  grouping_cols = "year",
  weights = "weight",
  method = c("Geometric", "Arithmetic", "Harmonic", "Median"),
  min_alpha = 1.11
)
}
\arguments{
\item{df}{A data frame containing the survey data}

\item{bands_df}{A table of bands, corresponding with the values in \code{band_col},
created using \code{\link[=create_bands_df]{create_bands_df()}}; fed into \code{\link[=compute_band_counts]{compute_band_counts()}}, and
\code{\link[=calculate_mcib]{calculate_mcib()}} via \code{\link[=calculate_riob]{calculate_riob()}}}

\item{band_col}{String; The name of the column in \code{df} with bands in, fed
into \code{\link[=compute_band_counts]{compute_band_counts()}}}

\item{grouping_cols}{String (default = "year"); Columns to group the counts
by, along with \code{band}, fed into \code{\link[=compute_band_counts]{compute_band_counts()}} (if no groups, set
to \code{NULL})}

\item{weights}{String (default = "weight"); The name of the column containing
the weights, fed into \code{\link[=compute_band_counts]{compute_band_counts()}}}

\item{method}{String; one of "Arithmetic", "Harmonic", "Geometric" or
"Median", fed into \code{\link[=calculate_rpme]{calculate_rpme()}} via \code{\link[=calculate_riob]{calculate_riob()}}}

\item{min_alpha}{Numeric; the minimum threshold for the alpha parameter of
the Pareto distribution, fed into \code{\link[=calculate_rpme]{calculate_rpme()}} via \code{\link[=calculate_riob]{calculate_riob()}}}
}
\description{
Combines Mean Constrained Integration Over Brackets and Robust Pareto
Midpoint Estimator to estimate the mean value of each band, in banded data.
Can be done by groups, e.g. year, or year and region, using the
\code{grouping_cols} argument.
}
\details{
Mean Constrained Integration Over Brackets:
\link{https://journals.sagepub.com/doi/10.1177/0081175018782579#sec-4}

We don't have a grand mean, so we only do the integration step.

Robust pareto midpoint estimator:
\link{https://ideas.repec.org/c/boc/bocode/s457962.html}
\link{https://arxiv.org/ftp/arxiv/papers/1402/1402.4061.pdf}

The default method is geometric and the default min_alpha value is 1.11,
which are the ideal options for the FBS.
These were chosen by calculating the Pareto Distribution using min_alpha
values from 0-4 in 0.01 increments for each of the four mean methods. The
method with the lowest errors was chosen as the default and of that method,
the alpha value with an error value closest to zero was chosen as the default
min_alpha value.
This will need to be updated if you are not using FBS data, or the FBS sample
size has significantly changed.
}
\seealso{
Other functions for estimating means of banded data: 
\code{\link{calculate_mcib}()},
\code{\link{calculate_riob}()},
\code{\link{calculate_rpme}()},
\code{\link{compute_band_counts}()},
\code{\link{create_bands_df}()}
}
\concept{functions for estimating means of banded data}
