% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix-suppression.R
\name{fix_suppression}
\alias{fix_suppression}
\title{Fix suppression in tables where totals can be used to calculate suppressed values}
\usage{
fix_suppression(
  df,
  suppression_type = c("mean_incomes", "outputs_costs", "details", "crop_stock_labour",
    "income_distributions", "balance_sheets", "liabilities", "total_diversified",
    "fbi_diversification", "diversified_outputs_incomes", "diversification_output_dist",
    "agricultural_transition", "ofi_median", "ofi_proportion", "machinery", "custom"),
  suppression_level = c("values", "sample_size"),
  sample_size_col = "sample_size",
  custom_groups = NULL
)
}
\arguments{
\item{df}{Data to fix}

\item{suppression_type}{Which table you are fixing}
}
\description{
This function finds which figures need to be secondarily suppress and
overwrites their sample size to zero - any subsequent suppression code will
then suppress these figures.
}
