% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix-suppression.R
\name{fix_suppression}
\alias{fix_suppression}
\title{Fix suppression in tables where totals can be used to calculate suppressed values}
\usage{
fix_suppression(
  df,
  groups,
  sample_size_col = "sample_size",
  export_path = "",
  file_name = "",
  pivot_col = NULL,
  save_excel_file = FALSE
)
}
\arguments{
\item{df}{Data to fix}

\item{groups}{A list of grouping dimensions, e.g. \code{list("group", c("band", "grouping_factor"))}
would group by the \code{group} column in the first dimension and by the \code{band} and
\code{grouping_factor} columns in the second dimension}

\item{sample_size_col}{Default = \code{"sample_size"}; The name of the sample size column}

\item{export_path}{File path; where the checking files should be saved}

\item{file_name}{What should the checking file be called? Don't include file extension}

\item{pivot_col}{String; If you want the manual checking table to be pivoted, which
column should it be pivoted on?}

\item{save_excel_file, }{default is TRUE, set to FALSE when one does not wish to save an excel file}
}
\description{
This function finds which figures need to be secondarily suppressed and
overwrites their sample size to one - any subsequent suppression code will
then suppress these figures.
}
\details{
This will loop though the group, or list of groupings, until there are no
more figures which need to be suppressed.

Sometimes this function can oversuppress, therefore, it also outputs an excel
file for manual checking. This file has the unsuppressed data in twice so that
can manually suppress one while comparing it to the unsuppressed data. The file
also has conditional formatting - values between 1 and 4 are highlighted in red
and values between 5 and 14 are highlighted in yellow.

If doing suppression across multiple groups, it is sometimes useful to consider
the order of the groups to minimize oversuppression.
}
\author{
Farm Business Survey team (\href{mailto:fbs.queries@defra.gov.uk}{fbs.queries@defra.gov.uk})
}
