% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-address.R
\name{get_address}
\alias{get_address}
\title{Get full URL for file attachments or links on gov.uk}
\usage{
get_address(
  web_address,
  file_type = NULL,
  file_number = 1,
  search_term = NULL,
  find_csv_preview = F
)
}
\arguments{
\item{web_address}{Webpage to look for the file on}

\item{file_type}{File extension to search for (e.g. "xlsx", "ods"); default
= \code{NULL}, leave as NULL if searching for a link rather than a file}

\item{file_number}{If there are multiple files of the same type as
\code{file_type} on the page, which one to return}

\item{search_term}{Regex; Optional, used by \code{str_subset} to filter on link URLs}

\item{find_csv_preview}{Logical, default = \code{FALSE}; when searching for a CSV
on gov.uk, this will return the link to the CSV preview when set to TRUE, and
will return the link to the actual CSV when set to FALSE}
}
\description{
Get full URL for file attachments or links on gov.uk
}
\seealso{
Other commentary functions: 
\code{\link{compare_two_years}()},
\code{\link{create_list}()},
\code{\link{get_diff_in_words}()},
\code{\link{get_pc_in_words}()}
}
\concept{commentary functions}
