% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commentary-functions.R
\name{get_diff_in_words}
\alias{get_diff_in_words}
\title{Get differences in text for statistics commentary}
\usage{
get_diff_in_words(option_num, pc, pc_rnd, curr_rnd = NULL, points = FALSE)
}
\arguments{
\item{option_num}{Numeric; the text option to return}

\item{pc}{Column containing the proportional difference between the two
values you are comparing}

\item{pc_rnd}{The rounded values from the column defined in the \code{pc} argument}

\item{curr_rnd}{Column containing the rounded values for the most recent data}

\item{points}{Logical, default = \code{FALSE}; is the difference in percentage
points rather than \%?}
}
\description{
There are a six output options to choose from:
\enumerate{
\item Past tense, displays the percentage change as text as well as the
\code{curr_rnd} value, e.g. \code{"increased by 10\% to 100"}
\item Present tense, displays the percentage change as text as well as the
\code{curr_rnd} value, e.g. \code{"increasing by 10\% to 100"}
\item Reformats option 2 to show the \code{curr_rnd} value first, e.g. \code{"100, an increase of 10\%"}
\item Uses 'higher'/'lower' to show only the percentage change, e.g. \code{"10\% higher"};
no need to set the \code{curr_rnd} argument
\item Uses 'rise'/'fall' to show only the percentage change, e.g. \code{"a rise of 10\%"};
no need to set the \code{curr_rnd} argument
\item Uses 'rose'/'fell' to show only the percentage change, e.g. \code{"rose by 10\%"};
no need to set the \code{curr_rnd} argument
}
}
\examples{
get_diff_in_words(1, .1, "10\%", curr_rnd = "100")
get_diff_in_words(2, .1, "10\%", curr_rnd = "100")
get_diff_in_words(3, .1, "10\%", curr_rnd = "100")
get_diff_in_words(4, .1, "10\%")
get_diff_in_words(5, .1, "10\%")
get_diff_in_words(6, .1, "10\%")

}
\seealso{
Other commentary functions: 
\code{\link{compare_two_years}()},
\code{\link{create_list}()},
\code{\link{get_address}()},
\code{\link{get_pc_in_words}()}
}
\author{
Farm Business Survey team (\href{mailto:fbs.queries@defra.gov.uk}{fbs.queries@defra.gov.uk})
}
\concept{commentary functions}
