% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commentary-functions.R
\name{get_pc_in_words}
\alias{get_pc_in_words}
\title{Get percentage in text}
\usage{
get_pc_in_words(pc, pc_rnd)
}
\arguments{
\item{pc}{Numeric; The percentage as a proportion, e.g. \code{0.1} for a value of 10\%}

\item{pc_rnd}{String; What to show if \code{pc} isn't converted to text, e.g. \code{"10\%"}}
}
\description{
Converts percentages to text, e.g. \code{0.25} will be converted to \code{"a quarter"}
}
\seealso{
Other commentary functions: 
\code{\link{compare_two_years}()},
\code{\link{create_list}()},
\code{\link{get_address}()},
\code{\link{get_diff_in_words}()}
}
\author{
Farm Business Survey team (\href{mailto:fbs.queries@defra.gov.uk}{fbs.queries@defra.gov.uk})
}
\concept{commentary functions}
