% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrite-num-cols.R
\name{overwrite_num_cols}
\alias{overwrite_num_cols}
\title{Fixes column where numbers are stored as text}
\usage{
overwrite_num_cols(excel_wb, sheet, cols, rows, df)
}
\arguments{
\item{excel_wb}{Openxlsx workbook name}

\item{sheet}{Worksheet (either name as string, or location as numeric)}

\item{cols}{Vector of column numbers to be overwritten}

\item{rows}{Vector of row numbers to be overwritten}

\item{df}{Data frame containing the data from the relevant worksheet}
}
\value{
Updated workbook with modified columns
}
\description{
Overwrites selected columns and rows that contain both numeric and character
elements. Once this is run, you should add a new style to the modified
columns and rows so that they are displayed correctly in the output file. Use
\code{\link[openxlsx:addStyle]{openxlsx::addStyle()}} and \code{\link[=get_cell_style]{get_cell_style()}} to do this.
}
\details{
This function is adapted from \code{rapid.spreadsheets::overwrite_df()}, to work
with \link[aftables:aftables-package]{aftables::aftables}; for the full rapid.spreadsheets code, see
\link{https://github.com/RAPID-ONS/rapid.spreadsheets/blob/main/R/create_data_table_tab.R}.
}
\examples{
\dontrun{
library(openxlsx)
library(aftables)

set.seed(1)

# Create an aftable
cover_df <- list("Section" = c("Title", "Content"))

contents_df <- data.frame("Sheet name" = "Table",
                          "Sheet title" = "Example",
                          check.names = FALSE)

table_df <- data.frame(
  Category = LETTERS[1:10],
  "Suppressed" = c(1:4, "[c]", 6:9, "[x]"),
  "Commas" = round_with_commas(rnorm(10) * 1e5, "optimise"),
  check.names = FALSE
)

aftable <- create_aftable(
  tab_titles = c("Cover", "Contents", contents_df$`Sheet name`),
  sheet_types = c("cover", "contents", "tables"),
  sheet_titles = c("Cover", "Contents", "Table"),
  sources = c(rep(NA_character_, 2), "Source"),
  tables = list(cover_df, contents_df, table_df))

excel_wb <- generate_workbook(aftable)

# Check the file
# note the format errors on the table sheet
openXL(excel_wb)

# Fix the errors
overwrite_num_cols(excel_wb, sheet = 3, cols = 2:3,
                   rows = 5:14, df = table_df)

# Check the file again
# the errors are gone, but the commas have disappeared
openXL(excel_wb)

# Add styling
addStyle(excel_wb, sheet = 3, cols = 2:3,
         rows = 5:14, gridExpand = TRUE,
         style = get_cell_style("number", "body"))

# Check the file a final time
# formatting is back and errors are still gone
openXL(excel_wb)
}

}
