% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrite-num-cols.R
\name{overwrite_num_cols}
\alias{overwrite_num_cols}
\title{Fixes column where numbers are stored as text}
\usage{
overwrite_num_cols(excel_wb, sheet, cols, rows, df)
}
\arguments{
\item{excel_wb}{Openxlsx workbook name}

\item{sheet}{Worksheet (either name as string, or location as numeric)}

\item{cols}{Vector of column numbers to be overwritten}

\item{rows}{Vector of row numbers to be overwritten}

\item{df}{Data frame containing the data from the relevant worksheet}
}
\value{
Updated workbook with modified columns
}
\description{
Overwrites selected columns and rows that contain both numeric and character
elements. Once this is run, you should add a new style to the modified
columns and rows so that they are displayed correctly in the output file. Use
\code{\link[openxlsx:addStyle]{openxlsx::addStyle()}} and \code{\link[=get_cell_style]{get_cell_style()}} to do this.
}
