% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script-addins.R
\name{script_addins}
\alias{script_addins}
\alias{defra_script_addin}
\alias{defra_roxygen_addin}
\title{Addin to open new scripts with suggested templates}
\usage{
defra_script_addin()

defra_roxygen_addin()
}
\value{
A script will open in RStudio.
}
\description{
There are two addins that will open a new blank R script which
is populated with one of two suggested script headers. The script addin will
open with our suggested script header, which contains key meta-data to help
you document your script properly. The roxygen addin will add the roxygen
skeleton to properly document functions. Although roxygen comments are mostly
used for documenting funtions in packages, we suggest using them to document
any function you create, as this will inevetably make it easier for others to
pick up.
}
