% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script-addins.R
\name{script_addins}
\alias{script_addins}
\alias{defra_script_addin}
\alias{defra_roxygen_addin}
\title{Internal RStudio addins to open new scripts with suggested templates}
\usage{
defra_script_addin()

defra_roxygen_addin()
}
\value{
Opens a new R script in RStudio; returns \code{NULL} invisibly.
}
\description{
These functions are designed to be used only via the RStudio
Addins menu. They open a new R script pre-populated with either:
\itemize{
\item A suggested script header
\item A Roxygen documentation skeleton
}

They use \code{\link[rstudioapi:rstudio-documents]{rstudioapi::documentNew()}} to open a new script pane in RStudio.
Calling these functions directly will not work correctly outside the
RStudio IDE.
}
\details{
These addins are part of the DefraUtils RStudio integration and are
intended to promote consistent documentation practices across projects.
They are not designed to be used interactively from the console.
}
\note{
These functions are triggered through the \strong{RStudio Addins} menu and
should not be called manually.
}
\seealso{
\code{\link[rstudioapi:rstudio-documents]{rstudioapi::documentNew()}}
}
\author{
Josh Moatt
}
\keyword{internal}
