% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-console-prompt.R
\name{set_console_prompt}
\alias{set_console_prompt}
\title{Add my R console prompt.}
\usage{
set_console_prompt(scope)
}
\arguments{
\item{scope}{string. "user" sets the prompt globally, "project" sets it just
for the active project.}
}
\value{
Altered R prompt
}
\description{
Function to change the R console prompt. The default prompt is
">". This function will update the prompt so that if you are in a Git
project, the prompt will display the acitve branch, e.g. "\link{@main}>". If you
are not in a Git enabled branch, the prompt will revert to the default ">".
Prompts can be manually edited using \code{usethis::edit_r_profile()}. Has two
scopes, "user" will apply this generally to RStudio, creating a user R
profile file, "project" will create a project specific R profile file.
}
\examples{
\dontrun{
# set user prompt
set_console_prompt("user")

# set project prompt
set_console_prompt("project")
}
}
\author{
Josh Moatt
}
