% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-databricks-pat.R
\name{set_databricks_pat}
\alias{set_databricks_pat}
\title{Prompt user to set Databricks Personal Access Token (PAT)}
\usage{
set_databricks_pat()
}
\value{
A character string containing the entered PAT. Typically assigned to
an environment variable using \code{Sys.setenv(DATABRICKS_TOKEN = token)}.
}
\description{
Prompts the user to enter their Databricks Personal Access Token
(PAT) interactively. This avoids hardcoding sensitive credentials in
scripts, reducing the risk of accidentally committing them to version
control.

This function is useful when working on the Defra DASH platform with the
\code{brickster} package. It is not required on Posit Workbench, which handles
credentials internally, and does not work in Databricks notebooks.
}
\examples{
\dontrun{
# set PAT
token <- set_databricks_pat()

# set to environmental variables
Sys.setenv(DATABRICKS_TOKEN = token)
}

}
\author{
Josh Moatt
}
