% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-model-functions.R
\name{test_model_performance}
\alias{test_model_performance}
\title{Check the sensitivity and specificity of a svyglm model}
\usage{
test_model_performance(survey_model, save_path = NULL)
}
\arguments{
\item{survey_model}{A svyglm object}

\item{save_path}{The directory where the plot should be saved to}
}
\description{
Currently only takes models with a continuous or binary response variable;
could be adapted for models with categorical response variables. If
\code{save_path} is set, the plot will be saved as a PNG to the specified directory.
}
\seealso{
Tested on models created with \code{\link[survey:svyglm]{survey::svyglm()}}

Other survey modelling functions: 
\code{\link{normalise_values}()},
\code{\link{run_wald_test}()},
\code{\link{summarise_variables}()}
}
\concept{survey modelling functions}
