% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-log.R
\name{tidy_log}
\alias{tidy_log}
\title{Tidy up a log created by futile.logger}
\usage{
tidy_log(log_path, export_path = log_path, export = TRUE, delim = "|")
}
\arguments{
\item{log_path}{the file path where the futile.logger log is saved}

\item{export_path}{the path to write the tidy log to}

\item{export}{logical; if TRUE (default) will export the tidied log as a csv
file, otherwise will return it to the console}
}
\description{
The messages you have written using by futile logger should have the
following format: "message: detail". For example:
\code{flog.info("Rows in data: \%s", nrow(data))}
Ensure that there is only one colon+space in the message, as ": " is used to
split the columns. This function will separate the log into four columns:
message_type, timestamp, message, and detail. If more than on colon+space is
used, everything after the first colon will end up in the detail column. If
you have not included a colon+space in your log message the detail column
will be populated with NA.
}
