% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-log.R
\name{tidy_log}
\alias{tidy_log}
\title{Tidy up a log created by futile.logger}
\usage{
tidy_log(log_path, export_path = log_path, export = TRUE, delim = "|")
}
\arguments{
\item{log_path}{the file path where the futile logger log is saved}

\item{export_path}{the path to write the tidy log to; the default is the
current location of the log, i.e., will overwrite the current log}

\item{export}{logical; if \code{TRUE} (default) will export the tidied log as a csv
file, otherwise will return it to the console}

\item{delim}{the delimiter used in the futile logger log; the default, \code{"|"},
should match any futile logger log, but the option has been given in case
futile.logger changes}
}
\description{
The messages you have written using \link[futile.logger:futile.logger-package]{futile.logger::futile.logger} should have the
following format: "message: detail", e.g.
\code{flog.info("Rows in data: \%s", nrow(data))}.
Or they can be a message without detail, e.g.
\code{flog.warn("Data file not found")}.
}
\details{
Ensure that there is only one \code{": "} in the message, as this string is used
to split the columns. This function will separate the log into four columns:
\itemize{
\item \code{message_type}
\item \code{timestamp}
\item \code{message}
\item \code{detail}
}

If more than one \code{": "} is used, everything after the first colon will end up
in the detail column. If you have not included \code{": "} in your log message,
the detail column will be populated with \code{NA}.
}
\author{
Farm Business Survey team (\href{mailto:fbs.queries@defra.gov.uk}{fbs.queries@defra.gov.uk})
}
