% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uc-volume-get.R
\name{uc_volume_get}
\alias{uc_volume_get}
\title{Download a file from the DASH Unity Catalog}
\usage{
uc_volume_get(workspace, volume, token, out_file)
}
\arguments{
\item{workspace}{String. The full URL of the Databricks workspace.}

\item{volume}{String. The full path to the file in the Unity Catalog (e.g.,
"/Volumes/...").}

\item{token}{String. Your Databricks Personal Access Token (PAT).}

\item{out_file}{String. The local file path on the RStudio server where the
file should be saved}
}
\value{
No return value. The file is saved to the specified location on the
RStudio server.
}
\description{
Downloads a file from the Databricks Unity Catalog (UC) to the
RStudio server. This function wraps a call to \code{\link[httr:GET]{httr::GET()}} and was adapted
from the DASH playbook.

You must have a valid Databricks Personal Access Token (PAT) set up in your
Databricks account.

For uploading files to UC, see \code{\link[=uc_volume_put]{uc_volume_put()}}.
}
\examples{
\dontrun{
uc_volume_get(
  workspace = "https://adb-7422054397937474.14.azuredatabricks.net",
  volume = "/Volumes/prd_dash_lab/<path-to-file>/filename.csv",
  token = "dapid4b3d********************a687f9b",
  out_file = here::here("data", "filename.csv")
)
}

}
\seealso{
\code{\link[httr:GET]{httr::GET()}}, \code{\link[=uc_volume_put]{uc_volume_put()}}
}
\author{
Josh Moatt
}
