% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uc-volume-put.R
\name{uc_volume_put}
\alias{uc_volume_put}
\title{Write file to the DASH Unity Catalog}
\usage{
uc_volume_put(workspace, volume, token, file, folder)
}
\arguments{
\item{workspace}{String. The full URL of the Databricks workspace.}

\item{volume}{String. The path to the Databricks volume or folder where the
file should be saved.}

\item{token}{String. Your Databricks Personal Access Token (PAT).}

\item{file}{String. The name of the file to upload, including its extension.}

\item{folder}{String. The local path to the folder containing the file to
upload (excluding the file name).}
}
\value{
No return value. The file is uploaded to the specified location in
the Unity Catalog.
}
\description{
Uploads a local file to the DASH Unity Catalog (UC) using a PUT
request via the \code{httr2} package. This function is useful for saving files
from the RStudio server to Databricks volumes.

You must have a valid Databricks Personal Access Token (PAT) configured in
your Databricks account.

For downloading files from UC, see \code{\link[=uc_volume_get]{uc_volume_get()}}.
}
\examples{
\dontrun{
uc_volume_put(
  workspace = "https://adb-7422054397937474.14.azuredatabricks.net",
  volume = "/Volumes/prd_dash_lab/<path-to-file>",
  token = "dapid4b3d********************a687f9b",
  file = "filename.csv",
  folder = here::here("data")
)
}

}
\seealso{
\code{\link[httr2:request]{httr2::request()}}, \code{\link[=uc_volume_get]{uc_volume_get()}}
}
\author{
Josh Moatt (\href{mailto:Joshua.Moatt@defra.gov.uk}{Joshua.Moatt@defra.gov.uk})
}
