% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-defra-stats-infographic.R
\name{view_available_icons}
\alias{view_available_icons}
\title{View icons}
\usage{
view_available_icons()
}
\description{
Returns a graphic in the Plots pane of the available icons in
\code{\link[=create_defra_stats_infographic]{create_defra_stats_infographic()}}. If text is overlapping in the Plots pane,
click 'Zoom' and resize the window to suit.
}
\author{
Farm Business Survey team (\href{mailto:fbs.queries@defra.gov.uk}{fbs.queries@defra.gov.uk})
}
